# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "Api do gerenciamento de times. Usuário Admin para testes [ Email: admin@example.com | Senha: admin123 ] "
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/users/create": {
      "post": {
        "tags": ["Users"],
        "description": "Cria um usuário.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "user name"
                },
                "email": {
                  "example": "example@example.com"
                },
                "password": {
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário criado com sucesso!"
          },
          "400": {
            "description": "Dados inválidos."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/create-admin": {
      "post": {
        "tags": ["Users"],
        "description": "Cria um usuário administrador.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "user name"
                },
                "email": {
                  "example": "example@example.com"
                },
                "password": {
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/read": {
      "get": {
        "tags": ["Users"],
        "description": "Buscar usuários por página.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Máximo de usuários exibidos por página. Valores possíveis: 5, 10, 30."
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Número da página a ser retornada."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "ID do usuário."
                    },
                    "username": {
                      "type": "string",
                      "description": "Nome de usuário."
                    },
                    "email": {
                      "type": "string",
                      "description": "Email do usuário."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID inválido."
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/update/{id}": {
      "put": {
        "tags": ["Users"],
        "description": "Atualizar dados de um usuário.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "user name",
                  "required": false
                },
                "email": {
                  "example": "example@example.com",
                  "required": false
                },
                "password": {
                  "example": "password",
                  "required": false
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/delete/{id}": {
      "delete": {
        "tags": ["Users"],
        "description": "Deletar um usuário.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}